requirements.txt
Django
djangorestframework
boto3
django-storages
